refactor(ExtractUtil): 클래스 역할에 맞게 명칭 변경
Service 의 역할 보다는 Util 성격에 더 가까워 명칭을 변경했습니다.
