+++ src/main/java/calculator/controller/CalculatorProcess.java
@@ -1,7 +1,7 @@
 package calculator.controller;
 
 import calculator.service.CalculatorService;
-import calculator.service.ExtractService;
+import calculator.util.ExtractUtil;
 import calculator.util.NumberValidator;
 import calculator.view.InputView;
 import calculator.view.OutputView;
@@ -10,14 +10,14 @@ public class CalculatorProcess {
     private final OutputView outputView;
     private final InputView inputView;
     private final CalculatorService calculatorService;
-    private final ExtractService extractService;
+    private final ExtractUtil extractUtil;
     private final NumberValidator numberValidator;
 
     public CalculatorProcess() {
         this.outputView = new OutputView();
         this.inputView = new InputView();
         this.calculatorService = new CalculatorService();
-        this.extractService = new ExtractService();
+        this.extractUtil = new ExtractUtil();
         this.numberValidator = new NumberValidator();
     }
 
@@ -25,8 +25,8 @@ public class CalculatorProcess {
         outputView.printStartMessage();
         String readInput = inputView.readInput();
 
-        String customDelimiter = extractService.extractCustomDelimiter(readInput);
-        String positiveString = extractService.extractPositiveString(readInput, customDelimiter);
+        String customDelimiter = extractUtil.extractCustomDelimiter(readInput);
+        String positiveString = extractUtil.extractPositiveString(readInput, customDelimiter);
 
         char[] positiveCharacters = calculatorService.toCharArrayWithoutDelimiter(positiveString, customDelimiter);
         numberValidator.validateNoDigits(positiveCharacters); // 숫자 변환 가능한 문자 검사
